Set icons for node buttons. Moved other buttons into menubar
I think it's important to have the screen as tidy as possible,
it makes my heart happy. There are also now key shortcuts to
all the menu buttons so you can use the actions without opening
any menus.
Also made the program open at the map loading screen, as it's
what the user will be clicking on basically all times.